--- app/src/main/java/com/example/calorietrackerapp/FoodItem.kt
@@ -1,4 +1,0 @@
-data class FoodItem(
-    val name: String,
-    val calories: Double
-)
