+++ app/src/main/java/com/example/calorietrackerapp/model/EdamamResponse.kt
@@ -1,4 +1,4 @@
-package com.example.calorietrackerapp
+package com.example.calorietrackerapp.model
 
 data class EdamamResponse(
     val hints: List<Hint>
@@ -15,4 +15,9 @@ data class Food(
 
 data class Nutrients(
     val ENERC_KCAL: Double
+)
+
+data class FoodItem(
+    val name: String,
+    val calories: Double
 )
